refactor: remove redundant metadata tests, consolidate into test_toolkit_metadata
- Delete tests/test_validation_metadata.py (redundant subset of test_toolkit_metadata.py)
- Remove TestValidationNamespace from test_decorator.py (duplicates convention tests)
- Port unique cases to test_toolkit_metadata.py: request_model shorthand,
  all-param-types, undecorated-function metadata absence
- Delete stale refactor/remove-validation-metadata branch

Closes #154
