+++ tests/test_decorator.py
@@ -131,96 +131,3 @@ class TestConfigurationErrors:
             def handler(req: HttpRequest, body: UserModel) -> HttpResponse:  # noqa: F811
                 return HttpResponse("ok")
 
-
-class TestValidationNamespace:
-    """Tests for metadata exposure on decorated functions."""
-
-    def test_metadata_attached_with_body(self) -> None:
-        """Decorated function exposes body model in metadata."""
-
-        @validate_http(body=UserModel)
-        def handler(req: HttpRequest, body: UserModel) -> HttpResponse:
-            return HttpResponse("ok")
-
-        meta = getattr(handler, "_azure_functions_metadata", None)
-        assert meta is not None
-        assert "validation" in meta
-        validation = meta["validation"]
-        assert validation["body"] is UserModel
-        assert validation["query"] is None
-        assert validation["path"] is None
-        assert validation["headers"] is None
-        assert validation["response_model"] is None
-
-    def test_metadata_attached_with_all_params(self) -> None:
-        """Decorated function exposes all configured models."""
-
-        @validate_http(
-            body=UserModel,
-            query=QueryModel,
-            path=PathModel,
-            headers=HeaderModel,
-            response_model=UserModel,
-        )
-        def handler(
-            req: HttpRequest,
-            body: UserModel,
-            query: QueryModel,
-            path: PathModel,
-            headers: HeaderModel,
-        ) -> HttpResponse:
-            return HttpResponse("ok")
-
-        meta = getattr(handler, "_azure_functions_metadata", None)
-        assert meta is not None
-        validation = meta["validation"]
-        assert validation["body"] is UserModel
-        assert validation["query"] is QueryModel
-        assert validation["path"] is PathModel
-        assert validation["headers"] is HeaderModel
-        assert validation["response_model"] is UserModel
-
-    def test_metadata_with_request_model_shorthand(self) -> None:
-        """request_model shorthand maps to body in metadata."""
-
-        @validate_http(request_model=UserModel)
-        def handler(req: HttpRequest, body: UserModel) -> HttpResponse:
-            return HttpResponse("ok")
-
-        meta = getattr(handler, "_azure_functions_metadata", None)
-        assert meta is not None
-        assert meta["validation"]["body"] is UserModel
-
-    def test_metadata_none_on_undecorated_function(self) -> None:
-        """Non-decorated function has no metadata."""
-
-        def handler(req: HttpRequest) -> HttpResponse:
-            return HttpResponse("ok")
-
-        assert getattr(handler, "_azure_functions_metadata", None) is None
-
-    def test_metadata_with_response_model_only(self) -> None:
-        """Metadata correctly captures response_model without request models."""
-
-        @validate_http(response_model=UserModel)
-        def handler(req: HttpRequest) -> dict[str, object]:
-            return {"name": "Alice", "age": 30}
-
-        meta = getattr(handler, "_azure_functions_metadata", None)
-        assert meta is not None
-        validation = meta["validation"]
-        assert validation["body"] is None
-        assert validation["response_model"] is UserModel
-
-    def test_metadata_on_async_handler(self) -> None:
-        """Async handlers also expose metadata."""
-
-        @validate_http(body=UserModel, response_model=UserModel)
-        async def handler(req: HttpRequest, body: UserModel) -> dict[str, object]:
-            return {"name": body.name, "age": body.age}
-
-        meta = getattr(handler, "_azure_functions_metadata", None)
-        assert meta is not None
-        validation = meta["validation"]
-        assert validation["body"] is UserModel
-        assert validation["response_model"] is UserModel
